Fix author insertion and comma rendering
Jeep had a bug where author insertion was not being performed correctly
(maintainer error that I blame on double negatives).

There was another bug with the way that the roff escape sequence, `\c`,
was being added to lines that caused commas (and trailing punctuation
generally) to not render correctly. A revised algorithm now avoids
adding the escape sequence if the previous line ends in trailing
punctuation.

This fixes issue #2 on GitHub.